Organize students section into parallel routes
