+++ src/app/course/[id]/@students/layout.tsx
@@ -1,26 +1,34 @@
 'use client';
 
-import { useEffect, useState } from "react";
+import { ReactNode, useEffect, useState } from "react";
 const store = require('store');
 
 import { styled } from "@mui/joy/styles";
-import Box from "@mui/joy/Box";
 import Stack from "@mui/joy/Stack";
+import Box from "@mui/joy/Box";
 
 import ToggleButtonGroup from '@mui/joy/ToggleButtonGroup';
 import IconButton from '@mui/joy/IconButton';
 import SvgIcon, { SvgIconProps } from '@mui/joy/SvgIcon';
 
-import AddStudentModal from "@/app/course/[id]/Students/AddStudentModal";
-import StudentsGallery from "@/app/course/[id]/Students/StudentsGallery";
-import StudentsTable from "@/app/course/[id]/Students/StudentsTable";
+import AddStudentModal from "@/app/course/[id]/@students/AddStudentModal";
 
 import AppsIcon from '../../../../../public/icons/apps.svg';
 import AppsFilledIcon from '../../../../../public/icons/appsFilled.svg';
 import TableIcon from '../../../../../public/icons/table.svg';
 import TableFilledIcon from '../../../../../public/icons/tableFilled.svg';
 
-const Students = ({ courseId }: { courseId: string }) => {
+const Layout = ({
+  children,
+  gallery,
+  table,
+  params,
+}: {
+  children: ReactNode,
+  gallery: ReactNode,
+  table: ReactNode,
+  params: { id: string },
+}) => {
   const [tabValue, setTabValue] = useState<string | null>('gallery');
 
   const storeVal = store.get('studentsView');
@@ -52,36 +60,35 @@ const Students = ({ courseId }: { courseId: string }) => {
   );
 
   return (
-    <Stack spacing={1}>
-      <Stack direction="row" justifyContent="space-between">
-        <Box>
-          <ToggleButtonGroup
-            value={tabValue}
-            onChange={(_, newValue) => {
-              store.set('studentsView', newValue);
-              setTabValue(newValue);
-            }}
-          >
-            <TabButton value="gallery">
-              <TabIcon component={tabValue === 'gallery' ? AppsFilledIcon : AppsIcon} />
-            </TabButton>
-            <TabButton value="table">
-              <TabIcon component={tabValue === 'table' ? TableFilledIcon : TableIcon} />
-            </TabButton>
-          </ToggleButtonGroup>
-        </Box>
-        <Box>
-          <AddStudentModal courseId={courseId} />
-        </Box>
+    <Stack gap={1}>
+      {children}
+      <Stack spacing={1}>
+        <Stack direction="row" justifyContent="space-between">
+          <Box>
+            <ToggleButtonGroup
+              value={tabValue}
+              onChange={(_, newValue) => {
+                store.set('studentsView', newValue);
+                setTabValue(newValue);
+              }}
+            >
+              <TabButton value="gallery">
+                <TabIcon component={tabValue === 'gallery' ? AppsFilledIcon : AppsIcon} />
+              </TabButton>
+              <TabButton value="table">
+                <TabIcon component={tabValue === 'table' ? TableFilledIcon : TableIcon} />
+              </TabButton>
+            </ToggleButtonGroup>
+          </Box>
+          <Box>
+            <AddStudentModal courseId={params.id} />
+          </Box>
+        </Stack>
+        
+        {tabValue === 'gallery' ? gallery : table}
       </Stack>
-      
-      {tabValue === 'gallery' ? (
-        <StudentsGallery courseId={courseId} />
-      ) : (
-        <StudentsTable courseId={courseId} />
-      )}
     </Stack>
   );
 };
 
-export default Students;
+export default Layout;
