+++ src/app/course/[id]/@students/page.tsx
@@ -1,0 +1,11 @@
+import Box from "@mui/joy/Box";
+
+const Students = () => {
+  return (
+    <Box textAlign="center">
+      <h4>Enrolled students</h4>
+    </Box>
+  );
+};
+
+export default Students;
