--- src/app/course/[id]/Students/index.ts
@@ -1,1 +1,0 @@
-export * from './Students';
