+++ src/app/course/[id]/page.tsx
@@ -1,13 +1,10 @@
 'use client';
 
-import Box from "@mui/joy/Box";
-import Stack from "@mui/joy/Stack";
 import Skeleton from "@mui/joy/Skeleton";
 import { styled } from "@mui/joy/styles";
 import Typography from "@mui/joy/Typography";
 import Grid from "@mui/joy/Grid";
 
-import Students from "@/app/course/[id]/Students/Students";
 import useCourse from "@/app/course/[id]/useCourse";
 
 export default function CourseDetails({ params }: { params: { id: string }}) {
@@ -25,55 +22,47 @@ export default function CourseDetails({ params }: { params: { id: string }}) {
   });
 
   return (
-    <Stack spacing={4} sx={{ display: 'flex', justifyContent: 'center' }}>
-      <Grid container maxWidth="68rem" width="100%" alignSelf="center">
-        <Grid xs={6} md={3}>
-          <Attribute>
-            <h4>Course</h4>
-            <AttributeValue>
-              <Skeleton loading={isLoading}>
-                {isLoading ? 'Biology II' : data?.name}
-              </Skeleton>
-            </AttributeValue>
-          </Attribute>
-        </Grid>
-        <Grid xs={6} md={3} textAlign="center">
-          <Attribute>
-            <h4>Teacher</h4>
-            <AttributeValue>
-              <Skeleton loading={isLoading}>
-                {isLoading ? 'Mr. Black' : data?.teacher}
-              </Skeleton>
-            </AttributeValue>
-          </Attribute>
-        </Grid>
-        <Grid xs={6} md={3} textAlign="center">
-          <Attribute>
-            <h4>Classroom</h4>
-            <AttributeValue>
-              <Skeleton loading={isLoading}>
-                {isLoading ? '101' : data?.classroom}
-              </Skeleton>  
-            </AttributeValue>
-          </Attribute>
-        </Grid>
-        <Grid xs={6} md={3} textAlign="center">
-          <Attribute>
-            <h4>Capacity</h4>
-            <AttributeValue>
-              <Skeleton loading={isLoading}>
-                {isLoading ? '11/30' : `${data?.students?.length}/${data?.capacity}`}
-              </Skeleton>
-            </AttributeValue>
-          </Attribute>
-        </Grid>
+    <Grid container maxWidth="68rem" width="100%" alignSelf="center">
+      <Grid xs={6} md={3}>
+        <Attribute>
+          <h4>Course</h4>
+          <AttributeValue>
+            <Skeleton loading={isLoading}>
+              {isLoading ? 'Biology II' : data?.name}
+            </Skeleton>
+          </AttributeValue>
+        </Attribute>
       </Grid>
-      <Stack gap={1}>
-        <Box textAlign="center">
-          <h4>Enrolled students</h4>
-        </Box>
-        <Students courseId={params.id} />
-      </Stack>
-    </Stack>
+      <Grid xs={6} md={3} textAlign="center">
+        <Attribute>
+          <h4>Teacher</h4>
+          <AttributeValue>
+            <Skeleton loading={isLoading}>
+              {isLoading ? 'Mr. Black' : data?.teacher}
+            </Skeleton>
+          </AttributeValue>
+        </Attribute>
+      </Grid>
+      <Grid xs={6} md={3} textAlign="center">
+        <Attribute>
+          <h4>Classroom</h4>
+          <AttributeValue>
+            <Skeleton loading={isLoading}>
+              {isLoading ? '101' : data?.classroom}
+            </Skeleton>  
+          </AttributeValue>
+        </Attribute>
+      </Grid>
+      <Grid xs={6} md={3} textAlign="center">
+        <Attribute>
+          <h4>Capacity</h4>
+          <AttributeValue>
+            <Skeleton loading={isLoading}>
+              {isLoading ? '11/30' : `${data?.students?.length}/${data?.capacity}`}
+            </Skeleton>
+          </AttributeValue>
+        </Attribute>
+      </Grid>
+    </Grid>
   );
 };
